# src/api/src/spec/openapi.yaml
openapi: 3.0.1
info:
  title: Resume Service
  version: 1.0.0
...
servers:
  - url: "/api/v1"
    description: resume api
paths:
...
  /user/login:
    post:
      tags:
        - User
      summary: login User
      operationId: postLoginUser
      requestBody:
        required: true
        $ref: '#/components/requestBodies/UserLoginBody' 
      responses:
        "202":
          $ref: "#/components/responses/UserDetail"
  /user/register:
    post:
      tags:
        - User
      summary: Register new user
      operationId: postRegisterUser
      requestBody:
        required: true
        $ref: '#/components/requestBodies/UserRegisterBody' 

      responses:
        "201":
          $ref: "#/components/responses/UserDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /query:
    get:
      tags:
        - Query
      x-middleware:
        - loginCheck
        - emailVerifyCheck
      summary: list queries
      operationId: getQueryList
      responses:
        "200":
          description: returns list of submitted queries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryList"
    post:
      tags:
        - Query
      summary: create query
      operationId: createQuery
      requestBody:
        required: true
        $ref: '#/components/requestBodies/QueryBody'
      responses:
        "201":
          description: create new query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryDetail"
  /meetingStatus:
    get:
      parameters:
        - $ref: "#/components/parameters/MeetingTimestamp"
      tags:
        - Meeting
      x-middleware:
        - loginCheck
        - emailVerifyCheck
      summary: list system meetings status
      operationId: getMeetingStatus
      responses:
        "200":
          description: returns list of system's meeting status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingStatusList"
  /meeting:
    get:
      tags:
        - Meeting
      x-middleware:
        - loginCheck
        - emailVerifyCheck
      summary: list meetings
      operationId: getMeetingList
      responses:
        "200":
          description: returns list of meetings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingList"
    post:
      tags:
        - Meeting
      x-middleware:
        - loginCheck
        - emailVerifyCheck
      summary: create Meeting
      operationId: createMeeting
      requestBody:
        required: true
        $ref: '#/components/requestBodies/MeetingBody'
      responses:
        "201":
          description: create new Meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingDetail"
  /meeting/{meetingId}:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    get:
      tags:
        - Meeting
      x-middleware:
        - loginCheck
        - emailVerifyCheck
      summary: get meeting
      operationId: getMeeting
      responses:
        "200":
          description: returns single meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingDetail"
    patch:
      tags:
        - Meeting
      x-middleware:
        - loginCheck
        - emailVerifyCheck
      summary: modify meeting
      operationId: patchMeeting
      requestBody:
        required: true
        $ref: '#/components/requestBodies/MeetingModificationBody'
      responses:
        "202":
          description: modify existing meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingDetail"
    delete:
      tags:
        - Meeting
      x-middleware:
        - loginCheck
        - emailVerifyCheck
      summary: delete meeting
      operationId: deleteMeeting
      responses:
        "202":
          description: delete existing meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Empty"
components:
...
  requestBodies:
    UserLoginBody:
      description: login email & password
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/LoginBody"
    UserRegisterBody:
      description: register user
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/RegisterUserDetail"
...
    QueryBody:
      description: query post detail
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/QueryDetailRequired"
    MeetingBody:
      description: meeting payload for creation
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/MeetingDetailRequired"
    MeetingModificationBody:
      description: meeting modification request body
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/MeetingDetailBase"

#######################################
###########   responses  ##############
#######################################
  responses:
    BadRequest:
      description: bad request
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Empty"
...
    UserDetail:
      description: user detail
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/UserDetail"

#######################################
############   Schemas  ###############
#######################################
  schemas:
...
    Empty:
      type: object
...
    ResourceList:
      required:
        - docs
      type: object
      properties:
        count:
          type: integer
          example: 1
        docs:
          type: array
          items:
            type: object
    createdAt:
        type: string
        format: date-time
        maxLength: 25
    DateMeta:
      required:
        - createdAt
        - modifiedAt
      type: object
      properties:
        createdAt:
          allOf:
            - $ref: "#/components/schemas/createdAt"
        modifiedAt:
          type: string
          format: date-time
          maxLength: 25
#######################################
##########  User Schemas  #############
#######################################
    LoginBody:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          maxLength: 256
        password:
          type: string
          maxLength: 25
    RegisterUserDetail:
      allOf:
        - $ref: "#/components/schemas/UserDetailBase"
      required:
        - password
      properties:
        password:
          type: string
          maxLength: 25
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/UserDetailBase"
        - $ref: "#/components/schemas/DateMeta"
      required:
        - id
        - emailVerified
        - mobileVerified
      properties:
        id:
          type: string
          example: "145664568768"
        emailVerified:
          type: boolean
        mobileVerified:
          type: boolean
    UserDetailBase:
      type: object
      required:
        - email
        - firstname
        - gender
        - mobile
        - dob
      properties:
        email:
          type: string
          maxLength: 256
        firstname:
          type: string
          maxLength: 100
        gender:
          type: string
          maxLength: 50
        lastname:
          type: string
          maxLength: 100
        dob:
          type: string
          format: date-time
          maxLength: 25
        mobile:
          type: string
          maxLength: 13
...
    QueryList:
      type: object
      allOf:
        - $ref: '#/components/schemas/ResourceList'
        - type: object
          properties:
            docs:
              type: array
              items:
                $ref: '#/components/schemas/QueryDetail'
    QueryDetail:
      allOf:
        - $ref: "#/components/schemas/QueryDetailRequired"
      type: object
      required:
        - id
        - createdAt
      properties:
        id:
          type: string
          example: "145664568768"
        createdAt:
          allOf:
            - $ref: "#/components/schemas/createdAt"
    QueryDetailRequired:
      allOf:
        - $ref: "#/components/schemas/QueryDetailBase"
      type: object
      required:
        - name
        - subject
        - mobile
        - email
    QueryDetailBase:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        subject:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 500
        email:
          type: string
          maxLength: 256
        mobile:
          type: string
          maxLength: 13

#######################################
############   Meetings  ##############
#######################################
    MeetingList:
      type: object
      allOf:
        - $ref: '#/components/schemas/ResourceList'
        - type: object
          properties:
            docs:
              type: array
              items:
                $ref: '#/components/schemas/MeetingDetail'
    MeetingDetail:
      allOf:
        - $ref: "#/components/schemas/MeetingDetailRequired"
        - $ref: "#/components/schemas/DateMeta"
      type: object
      required:
        - id
      properties:
        id:
          type: string
          example: "145664568768"
    MeetingDetailRequired:
      allOf:
        - $ref: "#/components/schemas/MeetingDetailBase"
      type: object
      required:
        - end
        - start
    MeetingDetailBase:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        start:
          type: string
          format: date-time
          maxLength: 25
        end:
          type: string
          format: date-time
          maxLength: 25
        members:
          type: array
          minItems: 0
          maxItems: 100
          items:
            type: string
            format: email
            maxLength: 256
            example: bob@gmail.com
    MeetingStatusList:
      type: object
      allOf:
        - $ref: '#/components/schemas/ResourceList'
        - type: object
          properties:
            docs:
              type: array
              items:
                $ref: '#/components/schemas/MeetingStatus'
    MeetingStatus:
      type: object
      required:
        - start
        - end
      properties:
        start:
          type: string
          format: date-time
          maxLength: 25
        end:
          type: string
          format: date-time
          maxLength: 25
#######################################
############  Identifiers  ############
#######################################
    Id:
      type: string
      pattern: '^[a-f\d]{24}$'
...
    MeetingId:
      allOf:
        - $ref: "#/components/schemas/Id"
        - description: Id of the Education


#######################################
############  Parameters  #############
#######################################
  parameters:
...
    MeetingId:
      name: meetingId
      description: meeting identifier
      in: path
      required: true
      schema:
        allOf:
          - $ref: "#/components/schemas/MeetingId"
    MeetingTimestamp:
      name: timestamp
      description: timestamp
      in: query
      required: true
      schema:
        type: string
        maxLength: 25
        format: date-time
      example: '2022-12-10T10:31:53.327Z'